SES callback (#90)
There is currently no way for SES to report the status of message delivery back to the API. Based on code fragments and some of the code in AU code base it looks like there used to be a route that was called by SNS when an email was processed. It looks like the GDS code base took this out and put it into a lambda function that updates the database directly, or uses a second non-api queue to handle the insertion. I have restored the URL based on the AU code and added tests. This should close #63