Upstreamz (#91)
* Remove deprecated fields from orgs and services

Once the admin app has started using the new fields returned in
https://github.com/alphagov/notifications-api/pull/2539 these fields
won’t be needed any more.

* Slim down the /organisations response

At the moment this response returns a list of service IDs for hundreds
of organisations.

The admin app doesn’t use this information, but having to wait for it to
be serialized and sent across the network slows it down all the same.

* Add extra fields to create/update organisation

* add emails to send when an MOU is signed

* tell the user that clicks the button
* tell the notify team
* tell the person that the button was clicked on behalf of

there are two templates for the first case

* Bump utils to bring in changes to email HTML

Changes:
https://github.com/alphagov/notifications-utils/compare/33.1.0...33.2.1

* only send emails if a user signed the MOU

the agreement_signed field can also be edited by a platform admin - if
that happened we might not have details for who signed it, and even if
we did they shouldn't find out about, and we don't need an email since
we were the ones who clicked the button.

the `agreement_signed_by` field is only set when a user confirms that
they are signing the MOU on the admin page - not if a platform admin
modifies the page from the platform admin page

* send emails when MOU is signed

we build up one personalisation dict, and then pass it in to all the
different templates - so be careful editing things. also of note, we
check if the agreement_signed_on_behalf_of is set, and send a different
template with slightly different wording to the person who clicked the
confirm button.

* add tests

include dynamically loading the template contents from the alembic upgrade file itself

* Set org_type of a service when adding/updating org org type

The organisation_type of a service should match the organisation_type of
the service's organisation (if there is one). This changes
dao_update_organisation and dao_add_service_to_organisation to set the
organisation_type of any services when adding / updating an org.

* Delete unused code

* The `_should_record_notification_in_history_table` function stopped being
used in this commit: c23ae15f32c4673b9d5a40f338e9e1eb87190430
* `NOTIFICATIONS_ALERT` stopped being used in this commit: 5aa37f09b63e7d85eac9224ed0084a2501ed3e84

* Bump utils to 33.2.2

Changes:
- https://github.com/alphagov/notifications-utils/compare/33.2.1...33.2.2

* Fix VSCode fail