Revert "[#17] set uk banner to false by default"
Although we don't want to use UK branding, the variable "govuk_banner" can be used in the template to show the VA's logo, which is what the Canadians are doing.

This piece of logic seemingly controls the default branding for services with no email branding - that default being "show the overall company/government/department banner".

This reverts commit d1c10fa7dca29e4e979570c10f9bf2a4249594d4.