Define load tests for one off email
This commit seeks to define load tests to test the one off email
endpoint[1].

Prior to this commit one example load tests already exists. This effort
expands the one off email enpoint load test to cover attachment and link.

This choice is influenced by the already existing smoke tests for one off email.

Locust insists on have a task defined[2], which it uses to simulate the to run the task,
this can be weighted which yields probable selection based on the weight.

The developers understand, would be that of isolated tasks that would run on thier own greenlet(thread),
hence this has led to the definition of seperate class in order to achieve the aforementioned.

[1] <domain>/v2/notifications/email
[2] https://docs.locust.io/en/stable/writing-a-locustfile.html