Revert to use Locust weighting for tasks
This commit makes changes to NotifyApiUser class by
collapsing the nested classes ad thier methods into
NotifyApiUser.

This means the tasks weight have been defined to ensure
that the likelihood of tasks running is twice as likely.

Also, the template id was changed to use the pre-existing
template that defines `application_file`