Reverse task weight.
Based on the author understanding prior to this commit.

The weight appeared to suggested that the lower the number the higher
the weight.

However, this isn't the case as pointed out by @sastels.

This commit reverses the weight order of the tasks to give the highest
weight to calls made to `/v2/notifications/email` endpoint.