Define Docker file for load/performance testing
To get daily performance tests going, we have identified the need for a
dedicated Docker file for this purpose.

This commit defines a new Docker file from which a runtime
Docker container will be created for usage in an AWS lambda.

With are defined the following environment variables:

- LOAD_TEST_PHONE_NUMBER
- LOAD_TEST_EMAIL
- LOAD_TEST_DOMAIN
- LOAD_TEST_SMS_TEMPLATE_ID
- LOAD_TEST_BULK_EMAIL_TEMPLATE_ID
- LOAD_TEST_EMAIL_TEMPLATE_ID
- LOAD_TEST_EMAIL_WITH_ATTACHMENT_TEMPLATE_ID
- LOAD_TEST_EMAIL_WITH_LINK_TEMPLATE_ID
- TEST_AUTH_HEADER

The aforelisted are used in [1], and these have been assigned for staging
specific values.

Given the definition within the Docker file, these environment variables can
overridden when this container is instantiated.

The performance tests defined in [1], also reuse some code from test_smoke
hence its inclusion.

[1] tests-perf/locust/locust-notifications.py