Build/Deploy Perf. Docker Image via GH Flow to ECR
This commit defines the build, push and deploy steps for performance
tests Docker images going forward.

It is triggered only when changes are detected to test-perf and
tests_smoke folders/paths[1].

The locust performance tests resuse code define under tests_smoke.

Most of these lines of YAML code is adapted from [2] and follows
predefined steps that are very similar to the goals of this commit.

[1] https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#onpushpull_requestpaths
[2] .github/workflows/build_and_push.yml