Publish performance test results
Within this commit, we have defined a script that setups the
performance test environment executes the performance test an
publishes it to S3 bucket.

The run time duration has also been set to 5 minutes to discourage endless
run of the performance tests

The s3 bucket is managed with a terraform script as defined here [1] and
will be used to hold the performance test results per environment
(production or staging).

[1] github.com/cds-snc/notification-terraform/pull/317