Experimental: New Relic initialization to app handler
Following on with our move to AWS Lambda, we require the ability to
track the performance of API via New Relic. With the way things are
AWS Lambda does not invoke gunicorn, so New Relic isn't getting
initialised.

This commit simply moves the New Relic initialization code to
application handler function which is invoked by the AWS Lambda.