fix: limit ALTER ROLE migration to active database
By default an `ALTER ROLE` statement applies to all databases in a
server.  This was causing a race condition in the unit tests which run
in multiple worker processes, each with their own database in the
same PostgreSQL server.

Adding the `IN DATABASE` clause limits the `ALTER ROLE` to only
apply to the database the migration is running against and removes
the race condition.

This fix also removes the pytest logging configuration.