feat: add default for str_to_bool function
Update the str_to_bool function to no longer throw an
exception if a string value cannot be parsed to a bool. In
these cases, a default value is returned instead.

This is done to move the default value control into the
Python code so it can be controlled for both the
Kubernetes pods and Lambda API function.