Report exception after first SMS sending retry only
This is to reduce alarms fatigue with our errors.

Do not log the exception right away when a SMS failure
occurs but only after the first retry. That is because
most likely error that happens at this stage is the db
data replication did not propagate yet and could not find
the notification. By retrying immediately as it is setup
at the moment, it usually succeeds on second try. But
prior to this change, the logging statement would trigger
our alarms even though the 2nd attemps most likely succeed.

If there is other type of errors, an immediate retry following
another failure will then log the exception and raise the
alarms as usual.