Fix bug where template redaction for P/Ts is flaky (#1888)
After testing creating new templates on staging for P/T services, templates were not always auto-redacted with the error `Template not found for redaction.` in the logs. This is logged when we try to fetch the template by id, and we get a NoResultFound error: https://github.com/cds-snc/notification-api/blob/main/app/template/rest.py#L106

This indicated that the template has not been created in the db yet and that is causing the redaction to fail. Instead, we can just set the redact_personalisation property to True when we create the template, and that seems to work.