Encoding unicode in "friendly-from" field (#1990)
This PR changes the way we are encoding the "friendly from" field in the email headers, so that we can include non-ascii characters that show up in service names. The AWS documentation provided examples of how to do this. There are 2 steps:
- encode the utf-8 characters as base64
- replace the ascii-encoded "friendly from" field with "MIME encoded-word syntax" that includes the base64 encoded string
Here is a link to the AWS docs that I used: https://docs.aws.amazon.com/ses/latest/dg/send-email-raw.html#send-email-mime-encoding-headers