Improve the formatting of the support emails for PTs (#2429)
This PR improves the formatting of the support emails for the "PT service skip Freshdesk" feature while still using the existing _generate_ticket method.

This PR removes the product_id, 'priority, status, and tagsdata from the email.priorityandstatuswere being hardcoded in this case so those numbers weren't meaningful.product_idwill always be the same for the prod app. Andtags` are only meaningful to freshdesk. So I think it is safe to remove this data for this feature since it makes the email less confusing for the support team.