Add a new decorator to rate limit db calls and unit tests (#2469)
This PR adds a new decorator to rate limit calls to the DB. This is used to limit DB calls to update the api key last_used_timestamp to once every 10 seconds. This is because we are getting db locks and very long wait times for this update operation during our performance tests.